requirements.txt
mss==7.0.1
Pillow==9.3.0
pynput==1.7.6
